backend/requirements.txt
asgiref==3.8.1
boto3==1.34.109
botocore==1.34.109
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==5.0.4
django-cors-headers==4.3.1
django-countries==7.6.1
django-elasticsearch-dsl==8.0
django-extensions==3.2.3
django-storages==1.14.3
django-templated-mail==1.1.1
django-unfold==0.29.1
django-vote==2.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
elastic-transport==8.13.1
elasticsearch==8.14.0
elasticsearch-dsl==8.14.0
idna==3.7
jmespath==1.0.1
joblib==1.4.2
mysqlclient==2.2.4
numpy==2.0.0
oauthlib==3.2.2
pandas==2.2.2
pillow==10.3.0
pycparser==2.22
pydot==2.0.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
s3transfer==0.10.1
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
threadpoolctl==3.5.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
